+++ printing/lib/src/pdf_preview.dart
@@ -22,6 +22,7 @@ class PdfPreview extends StatefulWidget {
     this.allowSharing = true,
     this.maxPageWidth,
     this.canChangePageFormat = true,
+    this.canChangeOrientation = true,
     this.actions,
     this.pageFormats,
     this.onError,
@@ -54,6 +55,9 @@ class PdfPreview extends StatefulWidget {
 
   /// Add a drop-down menu to choose the page format
   final bool canChangePageFormat;
+
+  /// Add a switch to change the page orientation
+  final bool canChangeOrientation;
 
   /// Additionnal actions to add to the widget
   final List<PdfPreviewAction> actions;
@@ -94,6 +98,8 @@ class _PdfPreviewState extends State<PdfPreview> {
 
   PdfPageFormat pageFormat;
 
+  bool horizontal;
+
   PrintingInfo info = PrintingInfo.unavailable;
   bool infoLoaded = false;
 
@@ -116,6 +122,10 @@ class _PdfPreviewState extends State<PdfPreview> {
     'Letter': PdfPageFormat.letter,
   };
 
+  PdfPageFormat get computedPageFormat => horizontal != null
+      ? (horizontal ? pageFormat.landscape : pageFormat.portrait)
+      : pageFormat;
+
   Future<void> _raster() async {
     Uint8List _doc;
 
@@ -124,7 +134,7 @@ class _PdfPreviewState extends State<PdfPreview> {
     }
 
     try {
-      _doc = await widget.build(pageFormat);
+      _doc = await widget.build(computedPageFormat);
     } catch (e) {
       error = e;
       return;
@@ -173,7 +183,12 @@ class _PdfPreviewState extends State<PdfPreview> {
     if (cc == 'US' || cc == 'CA' || cc == 'MX') {
       pageFormat = widget.initialPageFormat ?? PdfPageFormat.letter;
     } else {
-      pageFormat = widget.initialPageFormat ?? PdfPageFormat.a4;
+      pageFormat = widget.initialPageFormat;
+    }
+
+    final _pageFormats = widget.pageFormats ?? defaultPageFormats;
+    if (!_pageFormats.containsValue(pageFormat)) {
+      pageFormat = _pageFormats.values.first;
     }
 
     super.initState();
@@ -213,7 +228,7 @@ class _PdfPreviewState extends State<PdfPreview> {
       final mq = MediaQuery.of(context);
       dpi = (min(mq.size.width - 16, widget.maxPageWidth ?? double.infinity)) *
           mq.devicePixelRatio /
-          pageFormat.width *
+          computedPageFormat.width *
           72;
 
       _raster();
@@ -374,12 +389,41 @@ class _PdfPreviewState extends State<PdfPreview> {
           ),
           onChanged: (PdfPageFormat _pageFormat) {
             setState(() {
-              pageFormat = _pageFormat;
-              _raster();
+              if (_pageFormat != null) {
+                pageFormat = _pageFormat;
+                _raster();
+              }
             });
           },
         ),
       );
+
+      if (widget.canChangeOrientation) {
+        horizontal ??= pageFormat.width > pageFormat.height;
+        final color = theme.accentIconTheme.color;
+        final disabledColor = color.withAlpha(120);
+        actions.add(
+          ToggleButtons(
+            renderBorder: false,
+            borderColor: disabledColor,
+            color: disabledColor,
+            selectedBorderColor: color,
+            selectedColor: color,
+            children: <Widget>[
+              Transform.rotate(
+                  angle: -pi / 2, child: const Icon(Icons.note_outlined)),
+              const Icon(Icons.note_outlined),
+            ],
+            onPressed: (int index) {
+              setState(() {
+                horizontal = index == 1;
+                _raster();
+              });
+            },
+            isSelected: <bool>[horizontal == false, horizontal == true],
+          ),
+        );
+      }
     }
 
     if (widget.actions != null) {
@@ -393,7 +437,7 @@ class _PdfPreviewState extends State<PdfPreview> {
                 : () => action.onPressed(
                       context,
                       widget.build,
-                      pageFormat,
+                      computedPageFormat,
                     ),
           ),
         );
@@ -429,10 +473,13 @@ class _PdfPreviewState extends State<PdfPreview> {
           Material(
             elevation: 4,
             color: theme.primaryColor,
-            child: SafeArea(
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
-                children: actions,
+            child: SizedBox(
+              width: double.infinity,
+              child: SafeArea(
+                child: Wrap(
+                  alignment: WrapAlignment.spaceAround,
+                  children: actions,
+                ),
               ),
             ),
           )
@@ -441,7 +488,7 @@ class _PdfPreviewState extends State<PdfPreview> {
   }
 
   Future<void> _print() async {
-    var format = pageFormat;
+    var format = computedPageFormat;
 
     if (!widget.canChangePageFormat && pages.isNotEmpty) {
       format = PdfPageFormat(
@@ -451,14 +498,20 @@ class _PdfPreviewState extends State<PdfPreview> {
       );
     }
 
-    final result = await Printing.layoutPdf(
-      onLayout: widget.build,
-      name: widget.pdfFileName ?? 'Document',
-      format: format,
-    );
+    try {
+      final result = await Printing.layoutPdf(
+        onLayout: widget.build,
+        name: widget.pdfFileName ?? 'Document',
+        format: format,
+      );
 
-    if (result && widget.onPrinted != null) {
-      widget.onPrinted(context);
+      if (result && widget.onPrinted != null) {
+        widget.onPrinted(context);
+      }
+    } catch (e) {
+      if (widget.onError != null) {
+        widget.onError(context);
+      }
     }
   }
 
@@ -472,7 +525,7 @@ class _PdfPreviewState extends State<PdfPreview> {
         referenceBox.localToGlobal(referenceBox.paintBounds.bottomRight);
     final bounds = Rect.fromPoints(topLeft, bottomRight);
 
-    final bytes = await widget.build(pageFormat);
+    final bytes = await widget.build(computedPageFormat);
     final result = await Printing.sharePdf(
       bytes: bytes,
       bounds: bounds,